Re-org the program representation layer of the compiler
