+++ src/program/export.rs
@@ -1,0 +1,9 @@
+#[derive(Clone, Debug)]
+pub enum Export {
+    // TODO: Add other export types over time
+    Function,
+    Const,
+    Type,
+    OpMap,
+    TypeOpMap,
+}
