+++ src/program/fnkind.rs
@@ -1,0 +1,9 @@
+use crate::parse;
+
+#[derive(Clone, Debug)]
+pub enum FnKind {
+    Normal(Vec<parse::Statement>),
+    Bind(String),
+    Derived,
+    DerivedVariadic,
+}
