+++ src/program/operatormapping.rs
@@ -1,0 +1,70 @@
+use super::Export;
+use super::Scope;
+use crate::parse;
+
+#[derive(Clone, Debug)]
+pub enum OperatorMapping {
+    Prefix {
+        level: i8,
+        functionname: String,
+        operatorname: String,
+    },
+    Infix {
+        level: i8,
+        functionname: String,
+        operatorname: String,
+    },
+    Postfix {
+        level: i8,
+        functionname: String,
+        operatorname: String,
+    },
+}
+
+impl OperatorMapping {
+    pub fn from_ast(
+        scope: &mut Scope,
+        operatormapping_ast: &parse::OperatorMapping,
+        is_export: bool,
+    ) -> Result<(), Box<dyn std::error::Error>> {
+        let opmap = match operatormapping_ast.fix {
+            parse::Fix::Prefix(_) => OperatorMapping::Prefix {
+                level: operatormapping_ast
+                    .opmap
+                    .get_opprecedence()
+                    .num
+                    .parse::<i8>()?,
+                functionname: operatormapping_ast.opmap.get_fntoop().fnname.clone(),
+                operatorname: operatormapping_ast.opmap.get_fntoop().operator.clone(),
+            },
+            parse::Fix::Infix(_) => OperatorMapping::Infix {
+                level: operatormapping_ast
+                    .opmap
+                    .get_opprecedence()
+                    .num
+                    .parse::<i8>()?,
+                functionname: operatormapping_ast.opmap.get_fntoop().fnname.clone(),
+                operatorname: operatormapping_ast.opmap.get_fntoop().operator.clone(),
+            },
+            parse::Fix::Postfix(_) => OperatorMapping::Postfix {
+                level: operatormapping_ast
+                    .opmap
+                    .get_opprecedence()
+                    .num
+                    .parse::<i8>()?,
+                functionname: operatormapping_ast.opmap.get_fntoop().fnname.clone(),
+                operatorname: operatormapping_ast.opmap.get_fntoop().operator.clone(),
+            },
+        };
+        let name = match &opmap {
+            OperatorMapping::Prefix { operatorname, .. } => operatorname.clone(),
+            OperatorMapping::Infix { operatorname, .. } => operatorname.clone(),
+            OperatorMapping::Postfix { operatorname, .. } => operatorname.clone(),
+        };
+        if is_export {
+            scope.exports.insert(name.clone(), Export::OpMap);
+        }
+        scope.operatormappings.insert(name, opmap);
+        Ok(())
+    }
+}
